+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/dto/response/ReservationDetailResponseDto.java
@@ -66,6 +66,6 @@ public record ReservationDetailResponseDto(
 	/**
 	 * 현재 예약 상태 ID
 	 */
-	Long statusId
+	Integer statusId
 ) {
 }
