+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/repository/AdminReservationStatusRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * * 예약 상태(ReservationStatus) 엔티티에 대한 데이터 접근을 담당
  * * 주로 예약 상태 이름(Name)을 기반으로 해당 상태 엔티티를 조회하는 데 사용
  */
-public interface AdminReservationStatusRepository extends JpaRepository<ReservationStatus, Long> {
+public interface AdminReservationStatusRepository extends JpaRepository<ReservationStatus, Integer> {
 	/**
 	 * 주어진 예약 상태 이름과 일치하는 ReservationStatus 엔티티를 조회
 	 *
